dev-vcs/git-remote-dropbox: New.
"git-remote-dropbox" permits an unlimited number of private Git repositories
supporting an unlimited number of users to be freely hosted on Dropbox. This
certainly calls into question the current profit model of such popular Git
hosts as GitHub and BitBucket. The problem, of course, is that I like GitHub.
I wouldn't want to be in their unenviable (and increasingly untenable)
economic position. At the least, the future is unlikely to resemble the present.