games-engines/seventhsense: 1.17.0 added.
This commit adds the most recent stable release of "Seventh Sense",
David Olsen's phenomenal back- and frontend for Joe and Ben Dever's
"Lone Wolf" franchise of interactive role-playing fiction. Thanks to
Joe's voluntary relinquishment of all prior copyright to Project Aon,
the first eighteen books of the "Lone Wolf" series are freely as in both
beer and speech playable through this ebuild.

This is for all the unvanquished childhood nostalgia.