dev-python/cadquery-{pywrap,ocp}: Improved x 2.
This commit is the next in a commit chain improving the previously added
unstable ebuilds for the most recent stable releases of both pywrap and
OCP, mandatory dependencies of CadQuery. In particular, this commit now
replaces the broken pywrap 0.7.4 tarball with a working pywrap git
commit known to behave as expected. Both remain non-working. *sigh*