dev-python/cadquery-{pywrap,ocp}: Improved x 2.
This commit is the next in a commit chain improving the previously added
unstable ebuilds for the most recent stable releases of both pywrap and
OCP, mandatory dependencies of CadQuery. In particular, this commit
dramatically improves the cmake compilation portion of OCP. Both remain
non-working, but are now nearly working. *sigh*