games-roguelike/cataclysm-dda: Fix #106 x 2.
This commit is the next in a commit chain attempting to resolve issue
#106 with the Cataclysm: DDA, kindly submitted by @jrrom the brave
Gentoo acolyte. Specifically, this commit forcibly prevents the most
recent stable release of C:DDA from compiling itself with the `-Werror`,
enabling C:DDA to be compiled with erroneous casts. (This is awful. This
is C:DDA.)