games-roguelike/cataclysm-dda: "lto" USE flag fix.
If `USE=lto` was enabled, this would pull in `sys-devel/llvm[gold]`,
which has been removed from ::gentoo in favor of
`sys-devel/binutils[gold]` and cause the ebuild to fail with
unresolvable dependencies.

At the same time, upstream has made use of the Gold linker optional by
way of the build flag `GOLD`.

Empirically, when comparing Clang+LTO builds using ld.bfd and ld.gold,
build times during the linking phase are the same within error margins.
This does not justify choosing the option. This commit disables the
upstream Makefile's preference for it, allowing the ebuild to set its
own LDFLAGS if desired.

The Gold linker has generally fallen out of favor because its patrons
have moved on to LLVM, and active development elsewhere has continued in
ld.lld and ld.mold. These linkers are both valid options for this ebuild
with or without LTO enabled, but should be left for the user to enable
by way of Portage environment controls; e.g.,

/etc/portage/env/mold:
```
LDFLAGS="${LDFLAGS} -fuse-ld=mold"
```

/etc/portage/package.env/cataclysm-dda:
```
games-roguelike/cataclysm-dda mold
```

The above will, assuming the user has installed `sys-devel/mold`, ask
the package to be linked using `ld.mold`. This also happens to offer an
enormous speedup when compiling with GCC, and decouples the choice of
linker from the choices for LTO and compiler.