games-roguelike/cataclysm-dda: fix for GCC 13.
cf. CleverRaven/Cataclysm-DDA#66195

This packages a patch from the upstream PR, which simply includes the
`cstdint` library in a header file. Due to changes in the way C++
Standard Library headers are organized in GCC 13, the build fails
without this include.

Note that current live builds and future release builds already include
the fix. This commit should be reverted when release "0.G" is
superseded and this backporting patch is no longer required.